+++ include/parakeet/PointPolar.h
@@ -2,9 +2,11 @@
 	Copyright 2021 OpenJAUS, LLC (dba MechaSpin). Subject to the MIT license.
 */
 
-#pragma once
+#ifndef PARAKEET_POINTPOLAR_H
+#define PARAKEET_POINTPOLAR_H
 
 #include <cstdint>
+#include "macros.h"
 
 namespace mechaspin
 {
@@ -14,27 +16,39 @@ class PointPolar
 {
     public:
         /// \brief Create a PointPolar which is responsible for holding on to a polar coordinate.
-        /// \param[in] range - Distance from the origin
+        /// \param[in] range_mm - Distance, in millimeters, from the origin
         /// \param[in] angle_deg - Polar angle in degree form
         /// \param[in] intensity - The intensity value of the point
         /// \returns A PointPolar object which holds the information given through the params.
-        PointPolar(double range, double angle_deg, std::uint16_t intensity);
+        PointPolar(double range_mm, double angle_deg, std::uint16_t intensity);
 
-        /// \brief Returns the distance from the origin
-        /// \returns Distance from the origin
-        double getRange() const;
+        /// \deprecated Replaced by getRange_mm()
+        /// \brief Returns the distance, in millimeters, from the origin
+        /// \returns Distance, in millimeters, from the origin
+        PARAKEET_DEPRECATED(double getRange() const);
+
+        /// \brief Returns the distance, in millimeters, from the origin
+        /// \returns Distance, in millimeters, from the origin
+        double getRange_mm() const;
+
+        /// \deprecated Replaced by getAngle_deg()
+        /// \brief Returns the polar angle in degree form
+        /// \returns Polar angle in degree form
+        PARAKEET_DEPRECATED(double getAngleInDegrees() const);
 
         /// \brief Returns the polar angle in degree form
         /// \returns Polar angle in degree form
-        double getAngleInDegrees() const;
+        double getAngle_deg() const;
 
         /// \brief Returns the intensity value of the point
         /// \returns Intensity value of the point
         std::uint16_t getIntensity() const;
     private:
-        double range;
+        double range_mm;
         double angle_deg;
         std::uint16_t intensity;
 };
 }
 }
+
+#endif