+++ src/parakeet/internal/SerialPortHelper.cpp
@@ -4,13 +4,19 @@
 
 #include <parakeet/internal/SerialPortHelper.h>
 
-#ifdef __linux__
+#if defined(__linux) || defined(linux) || defined(__linux__)
 namespace ioctl
 {
     #include <sys/ioctl.h>
 }
 #include <asm/termios.h>
 
+namespace mechaspin
+{
+namespace parakeet
+{
+namespace internal
+{
 void SerialPortHelper::setCustomBaudRate(int fileDescriptor, int baudRate)
 {
     //Set baud rate, using termios2 here as termios1 does not support custom baud rates.
@@ -23,5 +29,8 @@ void SerialPortHelper::setCustomBaudRate(int fileDescriptor, int baudRate)
     tty2.c_ospeed = baudRate;
 
     ioctl::ioctl(fileDescriptor, TCSETS2, &tty2);
+}
+}
+}
 }
 #endif